+++ spec/spec_helper.rb
@@ -1,5 +1,16 @@
 # frozen_string_literal: true
 
+require 'simplecov'
+require 'simplecov-lcov'
+
+SimpleCov::Formatter::LcovFormatter.config.report_with_single_file = true
+SimpleCov::Formatter::LcovFormatter.config.single_report_path = 'coverage/lcov.info'
+SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter.new([
+                                                                 SimpleCov::Formatter::HTMLFormatter,
+                                                                 SimpleCov::Formatter::LcovFormatter
+                                                               ])
+SimpleCov.start
+
 require 'rubocop'
 require 'rubocop/rspec/support'
 require 'rubocop-exception_messages'