+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Initial release with `ExceptionMessages/Casing` and `ExceptionMessages/Punctuation` cops - [@dblock](https://github.com/dblock).
 * Add Danger CI check that enforces a CHANGELOG entry on every pull request - [@dblock](https://github.com/dblock).
 * Add RELEASING.md - [@dblock](https://github.com/dblock).
+* Add test coverage reporting with SimpleCov and Coveralls - [@dblock](https://github.com/dblock).
 
 #### Fixes